backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-docx==1.1.0
weasyprint==60.1
django-filter==23.5
drf-yasg==1.21.7
celery==5.4.7